Fix tests for l23network
This change includes an update to rspec-puppet ~> 2.2 and adjusting the
two tests that fail under 2.2 but not 2.1. I believe these failures are
related to https://github.com/rodjek/rspec-puppet/issues/266

The get_network_role_properly was previously passing because the return
value was not be properly check. In the code we are returning eth0 as
the interface for the admin network role and not nil.

The get_transformation_property was previously passing because the
return value was not correctly checked. Based on the code, the value for
eth2 should have been 1024 while the value for eth3 would have been nil.
This change updates the eth2 check with 1024 and adds a test case for
eth3.

FUEL-Change-Id: I7b87d02d4c702dde68ecd7f4abcd304e686fa795
FUEL-Closes-Bug: #1509056

Close: #182