Set rings rx/tx to maximum that supports hardware
All recent versions of the chipset have 4096, but the driver doesn't
autodetect this. Increasing the number of descriptors can improve
performance dramatically on some hosts.

This is also the case with Intel 10GE NICs - the driver defaults to 256
descriptors each for Rx and Tx, even if the hardware supports 4096.

FUEL-Change-Id: I1e5029e793399ff42002f3d203c6ae3bd2608be2
FUEL-Closes-Bug: #1456723

Closes: #273